+++ packages/cli/plugin-data-loader/tests/fixtures/server/bundles/three-server-loaders/index.js
@@ -1,5 +1,4 @@
-const { Response } = require('@remix-run/node');
-
+/* eslint-disable no-undef */
 const loader0 = async () => {
   return {
     message: 'loader0',