+++ packages/server/core/tests/base/utils/warmup.test.ts
@@ -1,0 +1,7 @@
+import { warmup } from '../../../src/base/utils';
+
+describe('test utils.warmup', () => {
+  it('should run correctly when import files is not exists', () => {
+    warmup(['abc']);
+  });
+});