+++ packages/server/prod-server/tests/helper.ts
@@ -1,8 +1,0 @@
-export const createDoc = () => `<html>
-  <head>
-    <title>mock</title>
-  </head>
-  <body>
-    <div>hello</div>
-  </body>
-  </html>`;