+++ packages/toolkit/utils/src/universal/path.ts
@@ -1,0 +1,3 @@
+export const cutNameByHyphen = (s: string) => {
+  return s.split(/[-_]/)[0];
+};