+++ packages/cli/plugin-lint/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/plugin-lint
 
+## 2.48.5
+
+### Patch Changes
+
+- Updated dependencies [4ca9f4c]
+  - @modern-js/utils@2.48.5
+  - @modern-js/tsconfig@2.48.5
+
 ## 2.48.4
 
 ### Patch Changes