+++ packages/generator/generator-utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/generator-utils
 
+## 3.3.16
+
+### Patch Changes
+
+- Updated dependencies [4ca9f4c]
+  - @modern-js/utils@2.48.5
+  - @modern-js/plugin-i18n@2.48.5
+  - @modern-js/generator-common@3.3.16
+
 ## 3.3.15
 
 ### Patch Changes