+++ packages/runtime/plugin-router-v5/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @modern-js/plugin-router-v5
 
+## 2.48.5
+
+### Patch Changes
+
+- Updated dependencies [4ca9f4c]
+  - @modern-js/utils@2.48.5
+  - @modern-js/plugin@2.48.5
+  - @modern-js/runtime-utils@2.48.5
+  - @modern-js/types@2.48.5
+
 ## 2.48.4
 
 ### Patch Changes