+++ packages/server/server/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @modern-js/server
 
+## 2.48.5
+
+### Patch Changes
+
+- Updated dependencies [4ca9f4c]
+  - @modern-js/utils@2.48.5
+  - @modern-js/prod-server@2.48.5
+  - @modern-js/server-utils@2.48.5
+  - @modern-js/runtime-utils@2.48.5
+  - @modern-js/types@2.48.5
+
 ## 2.48.4
 
 ### Patch Changes